Add more buttons and phrases
This adds more buttons and then sets up a set of phrases that can be
used instead of a single response. Each button has 3 possible phrases
when clicked that are chosed from a random array.

The array is passed to the selectPhrase function and using the Int
`arc4random_uniform` selects an index and then that index is used to
select the phrase from the array.

I've also consolidated the controller code into one function that can be
called by each of the buttons.

In the future I'd like to have 1 function and a case statement for the
phrases passed to the button, but I'm not sure how yet to access the
list of buttons from the code.