Consolidate methods into a case statement
Instead of making a new method for each case statement I assigned each
of the buttons a tag. Then I made a case statement that is assigned to
`arrayOfPhrases` and returns the array for that button. Once we have the
array we can pass it to the controller for handling which selects the
random phrase from the provided list.